Split util into separate file
